Chaskis/UnitTests/CoreTests/Handlers/Disconnected/DisconnectedEventConfigTests.cs
//
//          Copyright Seth Hendrick 2016-2021.
// Distributed under the Boost Software License, Version 1.0.
//    (See accompanying file LICENSE_1_0.txt or copy at
//          http://www.boost.org/LICENSE_1_0.txt)
//

using Chaskis.Core;
using NUnit.Framework;
using SethCS.Exceptions;

namespace Chaskis.UnitTests.CoreTests.Handlers.Disconnected
{
    [TestFixture]
    public class DisconnectedEventConfigTests
    {
        // ---------------- Tests ----------------

        [Test]
        public void ValidateTest()
        {
            DisconnectedEventConfig config = new DisconnectedEventConfig
            {
                LineAction = null
            };
            Assert.Throws<ListedValidationException>( () => config.Validate() );

            config.LineAction = delegate ( DisconnectedEventArgs args )
            {
            };

            Assert.DoesNotThrow( () => config.Validate() );
        }

        [Test]
        public void CloneTest()
        {
            DisconnectedEventConfig config = new DisconnectedEventConfig();
            DisconnectedEventConfig clone = config.Clone();

            Assert.AreNotSame( config, clone );
        }
    }
}
